move appcatalog to new resource
